fix container restarts can fail & update readme

+++ ubuntu-base/entrypoint.sh
@@ -1,11 +1,14 @@
 #!/bin/bash
 set -e
 
-/container-setup/setup-tz.sh.x
-/container-setup/setup-apt.sh.x
-
-# remove the setuid enabled executables for security
-rm -f /container-setup/*
+# skip the initialization in case of contaienr restarts
+if [[ ! -f /tmp/CONTAINER_HAS_STARTED ]]; then
+    /container-setup/setup-tz.sh.x
+    /container-setup/setup-apt.sh.x
+    # remove the setuid enabled executables for security
+    rm -f /container-setup/*
+fi
+echo 1 > /tmp/CONTAINER_HAS_STARTED
 
 if [[ $EUID -ne 0 ]]; then
     exec "$@"
